fix(extension): keep graph search compact
