+++ .changeset/show-workspace-graph-title.md
@@ -2,6 +2,6 @@
 "@codegraphy-dev/extension": minor
 ---
 
-Show the current workspace name as the Graph view title and move graph search controls into a dedicated Search view above the graph.
+Show the current workspace name as the Graph view title and keep graph search controls in a compact strip above the graph.
 
-The Search view keeps the search field, match-case/whole-word/regex toggles, and filter controls available while graph data is loading or refreshing. The Graph view now has more room for the graph itself while continuing to react to the same search and filter state.
+The Graph view now shows the current workspace name in VS Code's view title, falling back to Graph when no workspace is open. Search, match-case/whole-word/regex toggles, and filter controls stay available while graph data is loading or refreshing without creating a separate blank sidebar pane.
