+++ packages/extension/src/extension/graphView/provider/webview/resolve.ts
@@ -5,26 +5,15 @@ import type { GraphViewProviderSidebarViewSource } from './sidebarViews';
 
 export interface GraphViewProviderWebviewResolveSource extends GraphViewProviderSidebarViewSource {
   _extensionUri: vscode.Uri;
-  _searchViewCompacted?: boolean;
   _getLocalResourceRoots(): vscode.Uri[];
   flushPendingWorkspaceRefresh?(): void;
 }
-
-const SEARCH_VIEW_DECREASE_STEPS = 24;
 
 function isTimelineWebviewView(webviewView: vscode.WebviewView): boolean {
   return webviewView.viewType === 'codegraphy.timelineView';
 }
 
-function isSearchWebviewView(webviewView: vscode.WebviewView): boolean {
-  return webviewView.viewType === 'codegraphy.searchView';
-}
-
 function getWebviewKind(webviewView: vscode.WebviewView): CodeGraphyWebviewKind {
-  if (isSearchWebviewView(webviewView)) {
-    return 'search';
-  }
-
   if (isTimelineWebviewView(webviewView)) {
     return 'timeline';
   }
@@ -38,11 +27,6 @@ function assignResolvedWebviewView(
   viewKind: CodeGraphyWebviewKind,
   workspaceTitle: string | undefined,
 ): void {
-  if (viewKind === 'search') {
-    source._searchView = webviewView;
-    return;
-  }
-
   if (viewKind === 'timeline') {
     source._timelineView = webviewView;
     return;
@@ -57,10 +41,6 @@ function clearResolvedWebviewView(
   webviewView: vscode.WebviewView,
   viewKind: CodeGraphyWebviewKind,
 ): void {
-  if (viewKind === 'search' && source._searchView === webviewView) {
-    source._searchView = undefined;
-  }
-
   if (viewKind === 'timeline' && source._timelineView === webviewView) {
     source._timelineView = undefined;
   }
@@ -78,26 +58,6 @@ function maybeFlushPendingWorkspaceRefresh(
   if (viewKind === 'graph' && webviewView.visible) {
     source.flushPendingWorkspaceRefresh?.();
   }
-}
-
-function compactSearchView(
-  source: GraphViewProviderWebviewResolveSource,
-  dependencies: Pick<GraphViewProviderWebviewMethodDependencies, 'executeCommand'>,
-  viewKind: CodeGraphyWebviewKind,
-): void {
-  if (viewKind !== 'search' || source._searchViewCompacted) {
-    return;
-  }
-
-  source._searchViewCompacted = true;
-  void (async () => {
-    await dependencies.executeCommand('workbench.action.openView', 'codegraphy.searchView');
-    await dependencies.executeCommand('workbench.action.focusSideBar');
-    for (let index = 0; index < SEARCH_VIEW_DECREASE_STEPS; index += 1) {
-      await dependencies.executeCommand('workbench.action.decreaseViewHeight');
-    }
-    await dependencies.executeCommand('workbench.action.openView', 'codegraphy.graphView');
-  })().catch(() => undefined);
 }
 
 export function resolveGraphViewProviderWebviewView(
@@ -138,6 +98,5 @@ export function resolveGraphViewProviderWebviewView(
       dependencies.executeCommand(command, key, value),
   } as never);
 
-  compactSearchView(source, dependencies, viewKind);
   maybeFlushPendingWorkspaceRefresh(source, webviewView, viewKind);
 }
