+++ packages/extension/src/extension/graphView/provider/webview/sidebarViews.ts
@@ -1,7 +1,6 @@
 import type * as vscode from 'vscode';
 
 export interface GraphViewProviderSidebarViewSource {
-  _searchView?: vscode.WebviewView;
   _view?: vscode.WebviewView;
   _timelineView?: vscode.WebviewView;
 }
@@ -9,7 +8,7 @@ export interface GraphViewProviderSidebarViewSource {
 export function getGraphViewProviderSidebarViews(
   source: GraphViewProviderSidebarViewSource,
 ): vscode.WebviewView[] {
-  return [source._searchView, source._view, source._timelineView].filter(
+  return [source._view, source._timelineView].filter(
     (view): view is vscode.WebviewView => view !== undefined,
   );
 }
