+++ packages/extension/src/extension/graphViewProvider.ts
@@ -15,7 +15,6 @@ export class GraphViewProvider
   extends GraphViewProviderRuntime
   implements vscode.WebviewViewProvider
 {
-  public static readonly searchViewType = 'codegraphy.searchView';
   public static readonly viewType = 'codegraphy.graphView';
   public static readonly timelineViewType = 'codegraphy.timelineView';
 }
