+++ packages/extension/src/webview/app/shell/view.tsx
@@ -1,7 +1,9 @@
 import React, { useEffect, useState } from 'react';
+import { postMessage } from '../../vscodeApi';
 import { useTheme } from '../../theme/useTheme';
 import { usePluginManager } from '../../pluginRuntime/useManager';
 import { useFilteredGraph } from '../../search/useFilteredGraph';
+import type { SearchOptions } from '../../components/searchBar/field/model';
 import { getNoDataHint } from './messages';
 import { setupMessageListener } from './messageListener';
 import { LoadingState, EmptyState } from './states';
@@ -18,6 +20,10 @@ import { useRulePromptHandlers } from '../rulePrompt/handlers';
 import { buildPendingFilterPatterns } from './filterPopover';
 import { useVisibleGraphStateResponse } from './visibleGraphResponse';
 import { ActiveFileBreadcrumb } from '../../components/activeFileBreadcrumb/view';
+import { SearchHeader } from './panel/search';
+import { useShellVisibleGraphs } from './visibleGraphs';
+import { getShellGraphCountState } from './counts';
+import { useFilterPopoverState } from './filterPopover';
 
 export default function App(): React.ReactElement {
   const { pluginHost, injectPluginAssets, resetPluginAssets } = usePluginManager();
@@ -30,6 +36,7 @@ export default function App(): React.ReactElement {
     legends,
     filterPatterns,
     pluginFilterPatterns,
+    pluginFilterGroups,
     disabledCustomFilterPatterns,
     disabledPluginFilterPatterns,
     showOrphans,
@@ -48,11 +55,20 @@ export default function App(): React.ReactElement {
     graphIndexProgress,
   } = useAppState();
   const {
+    setSearchQuery,
+    setSearchOptions,
     setActivePanel,
     setFilterPatterns,
+    setDisabledCustomFilterPatterns,
+    setDisabledPluginFilterPatterns,
   } = useAppActions();
   const setOptimisticUserLegends = useGraphStore((state) => state.setOptimisticUserLegends);
   const [rulePrompt, setRulePrompt] = useState<RulePromptState | null>(null);
+  const {
+    filterPopoverOpen,
+    handleFilterPopoverOpenChange,
+    pendingFilterPatterns,
+  } = useFilterPopoverState();
 
   const theme = useTheme();
   const { activeFilterPatterns, userLegendRules } = useFilterLegendInputs(
@@ -63,9 +79,21 @@ export default function App(): React.ReactElement {
     legends,
   );
   const effectiveShowOrphans = graphHasIndex ? showOrphans : true;
+  const { countBaseData, filterVisibleData } = useShellVisibleGraphs({
+    activeFilterPatterns,
+    edgeVisibility,
+    graphData,
+    graphEdgeTypes,
+    graphNodeTypes,
+    nodeVisibility,
+    searchOptions,
+    showOrphans: effectiveShowOrphans,
+  });
   const {
     coloredData,
+    filteredData,
     edgeDecorations: graphEdgeDecorations,
+    regexError,
   } = useFilteredGraph(
     graphData,
     searchQuery,
@@ -101,17 +129,35 @@ export default function App(): React.ReactElement {
   const displayGraphData = coloredData || graphData;
   useVisibleGraphStateResponse(displayGraphData);
 
-  if (isLoading) return <LoadingState />;
+  const { countState, countTotal, excludedCount } = graphData
+    ? getShellGraphCountState({
+      countBaseData,
+      filterVisibleData,
+      filteredData,
+      graphData,
+      regexError,
+      searchQuery,
+    })
+    : {
+      countState: { label: null },
+      countTotal: 0,
+      excludedCount: 0,
+    };
+  const commitSearchState = (query: string, options: SearchOptions) => {
+    postMessage({
+      type: 'UPDATE_SEARCH_STATE',
+      payload: { query, options },
+    });
+  };
+  const handleSearchQueryChange = (query: string) => {
+    setSearchQuery(query);
+    commitSearchState(query, searchOptions);
+  };
+  const handleSearchOptionsChange = (options: SearchOptions) => {
+    setSearchOptions(options);
+    commitSearchState(searchQuery, options);
+  };
 
-  if (!graphData) {
-    return <EmptyState hint={getNoDataHint(graphData, showOrphans, depthMode, timelineActive)} />;
-  }
-
-  const loadedDisplayGraphData = displayGraphData ?? graphData;
-  const graphStatsLabel = buildGraphStatsLabel(
-    loadedDisplayGraphData.nodes.length,
-    loadedDisplayGraphData.edges.length,
-  );
   const closeActivePanel = () => setActivePanel('none');
   const openFilterPromptWithPatterns = (patterns: string[]) => {
     const [pattern] = buildPendingFilterPatterns(patterns);
@@ -119,15 +165,23 @@ export default function App(): React.ReactElement {
       openFilterPrompt(pattern);
     }
   };
+  const graphContent = (() => {
+    if (isLoading) {
+      return <LoadingState fullScreen={false} />;
+    }
 
-  return (
-    <div className="relative w-full h-screen flex flex-col">
-      {activeFilePath && (
-        <div className="flex-shrink-0 border-b border-border px-2 py-1">
-          <ActiveFileBreadcrumb filePath={activeFilePath} />
-        </div>
-      )}
-      <div className="flex-1 min-h-0 relative">
+    if (!graphData) {
+      return <EmptyState fullScreen={false} hint={getNoDataHint(graphData, showOrphans, depthMode, timelineActive)} />;
+    }
+
+    const loadedDisplayGraphData = displayGraphData ?? graphData;
+    const graphStatsLabel = buildGraphStatsLabel(
+      loadedDisplayGraphData.nodes.length,
+      loadedDisplayGraphData.edges.length,
+    );
+
+    return (
+      <>
         <GraphSurface
           graphData={graphData}
           coloredData={coloredData}
@@ -155,6 +209,44 @@ export default function App(): React.ReactElement {
           onClose={closeRulePrompt}
           onSubmit={handleRulePromptSubmit}
         />
+      </>
+    );
+  })();
+
+  return (
+    <div className="relative w-full h-screen flex flex-col">
+      <SearchHeader
+        searchQuery={searchQuery}
+        searchOptions={searchOptions}
+        resultCount={filteredData?.nodes.length}
+        totalCount={countTotal}
+        activeFilePath={null}
+        countLabel={countState.label}
+        filterPopover={{
+          customPatterns: filterPatterns,
+          disabledCustomPatterns: disabledCustomFilterPatterns,
+          disabledPluginPatterns: disabledPluginFilterPatterns,
+          excludedCount,
+          onDisabledCustomPatternsChange: setDisabledCustomFilterPatterns,
+          onDisabledPluginPatternsChange: setDisabledPluginFilterPatterns,
+          onOpenChange: handleFilterPopoverOpenChange,
+          onPatternsChange: setFilterPatterns,
+          open: filterPopoverOpen,
+          pendingPatterns: pendingFilterPatterns,
+          pluginGroups: pluginFilterGroups,
+          pluginPatterns: pluginFilterPatterns,
+        }}
+        regexError={regexError}
+        onSearchQueryChange={handleSearchQueryChange}
+        onSearchOptionsChange={handleSearchOptionsChange}
+      />
+      {activeFilePath && (
+        <div className="flex-shrink-0 border-b border-border px-2 py-1">
+          <ActiveFileBreadcrumb filePath={activeFilePath} />
+        </div>
+      )}
+      <div className="flex-1 min-h-0 relative">
+        {graphContent}
       </div>
     </div>
   );
