+++ packages/extension/tests/webview/main.test.tsx
@@ -28,12 +28,6 @@ vi.mock('../../src/webview/app/timeline/view', () => ({
   },
 }));
 
-vi.mock('../../src/webview/app/search/view', () => ({
-  default: function SearchApp() {
-    return null;
-  },
-}));
-
 vi.mock('../../src/webview/index.css', () => ({}));
 
 vi.mock('../../src/webview/vscodeApi', () => ({
@@ -90,7 +84,7 @@ describe('main', () => {
     expect(rootElement.props.children.type.name).toBe('TimelineApp');
   });
 
-  it('renders the search shell when the host marks the view as search', async () => {
+  it('renders the graph shell when the host provides an unknown view kind', async () => {
     const container = document.createElement('div');
     document.body.dataset.codegraphyView = 'search';
     vi.spyOn(document, 'getElementById').mockReturnValue(container);
@@ -100,7 +94,7 @@ describe('main', () => {
     const rootElement = mocks.render.mock.calls[0]?.[0] as {
       props: { children: { type: { name: string } } };
     };
-    expect(rootElement.props.children.type.name).toBe('SearchApp');
+    expect(rootElement.props.children.type.name).toBe('GraphApp');
   });
 
   it('skips root creation when the root element is missing', async () => {
